Update README.md
This Analysis examines Airbnb reviews to gain a better understanding of consumers' attitudes and expectations, providing both qualitative and quantitative data on travel behaviors in london. The main goal is to help Airbnb and its hosts improve their services, leading in increased customer happiness and better business decisions. The emphasis is on delivering individualized, comfortable accommodations.
For this Analysis we have used Sentiment Analysis, Correlation Anlysis and Topic Modelling.